requirements.txt
# LeadVille Impact Bridge Requirements
# Production BLE-based impact sensor system

# Core BLE and async dependencies
bleak>=0.20.0              # Bluetooth Low Energy library for Python
asyncio>=3.4.3             # Asynchronous I/O support

# Data processing and analysis
numpy>=1.21.0              # Numerical computing library
scipy>=1.7.0               # Scientific computing library
pandas>=1.3.0              # Data manipulation and analysis

# Optional: MQTT for remote monitoring (if needed)
# asyncio-mqtt>=0.11.0     # MQTT client for async applications

# Optional: Advanced data visualization (if needed)  
# matplotlib>=3.5.0        # Plotting library
# seaborn>=0.11.0          # Statistical data visualization

# Database and persistence
sqlalchemy>=2.0.0          # SQL toolkit and ORM
alembic>=1.12.0            # Database migration tool

# MQTT message bus
paho-mqtt>=1.6.0           # MQTT client library

# Web frameworks and API
fastapi>=0.104.0           # Modern, fast web framework for building APIs
uvicorn>=0.24.0            # ASGI server for FastAPI
websockets>=11.0           # WebSocket support for FastAPI

# Authentication and security
PyJWT>=2.8.0               # JSON Web Token implementation
bcrypt>=4.0.0              # Password hashing library
flask>=2.3.0               # Web framework for API endpoints (legacy)
flask-cors>=4.0.0          # CORS support for API (legacy)

# Configuration management  
pyyaml>=6.0                # YAML configuration files
# toml>=0.10.2             # TOML configuration files

# Development and testing dependencies (install separately)
# pytest>=6.0              # Testing framework
# pytest-asyncio>=0.18.0   # Async testing support
# black>=22.0              # Code formatter
# flake8>=4.0              # Linting tool
# mypy>=0.950              # Static type checker